gh-1436 - moved graph frames handler into the spark-library
It used to be in the spark-accumulo-library, but it does not required accumulo, it would work for any Gaffer Store.
