+++ library/spark/spark-accumulo-library/src/main/resources/sparkAccumuloOperationsDeclarations.json
@@ -57,7 +57,7 @@
     {
       "operation": "uk.gov.gchq.gaffer.spark.operation.graphframe.GetGraphFrameOfElements",
       "handler": {
-        "class": "uk.gov.gchq.gaffer.sparkaccumulo.operation.handler.graphframe.GetGraphFrameOfElementsHandler"
+        "class": "uk.gov.gchq.gaffer.spark.operation.handler.graphframe.GetGraphFrameOfElementsHandler"
       }
     }
   ]
