+++ library/spark/spark-library/src/main/java/uk/gov/gchq/gaffer/spark/operation/handler/graphframe/GetGraphFrameOfElementsHandler.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.gchq.gaffer.sparkaccumulo.operation.handler.graphframe;
+package uk.gov.gchq.gaffer.spark.operation.handler.graphframe;
 
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
@@ -23,13 +23,12 @@ import org.apache.spark.sql.expressions.Window;
 import org.apache.spark.sql.functions;
 import org.graphframes.GraphFrame;
 
-import uk.gov.gchq.gaffer.accumulostore.AccumuloStore;
 import uk.gov.gchq.gaffer.operation.OperationException;
 import uk.gov.gchq.gaffer.spark.SparkContextUtil;
 import uk.gov.gchq.gaffer.spark.operation.dataframe.GetDataFrameOfElements;
 import uk.gov.gchq.gaffer.spark.operation.dataframe.converter.schema.SchemaToStructTypeConverter;
 import uk.gov.gchq.gaffer.spark.operation.graphframe.GetGraphFrameOfElements;
-import uk.gov.gchq.gaffer.sparkaccumulo.operation.utils.scala.DataFrameUtil;
+import uk.gov.gchq.gaffer.spark.utils.scala.DataFrameUtil;
 import uk.gov.gchq.gaffer.store.Context;
 import uk.gov.gchq.gaffer.store.Store;
 import uk.gov.gchq.gaffer.store.operation.handler.OutputOperationHandler;
@@ -38,26 +37,19 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
+ * <p>
  * A {@code GetGraphFrameOfElementsHandler} handles {@link GetGraphFrameOfElements}
  * operations.
+ * </p>
  * <p>
- * The implementation found here is very similar to the {@link uk.gov.gchq.gaffer.sparkaccumulo.operation.handler.dataframe.GetDataFrameOfElementsHandler}
- * implementation. The main difference is that the resulting {@link Dataset} of
- * elements are split into two {@link Dataset}s based on the groups provided in
- * the {@link uk.gov.gchq.gaffer.data.elementdefinition.view.View}.
- *
- * @see uk.gov.gchq.gaffer.sparkaccumulo.operation.handler.dataframe.GetDataFrameOfElementsHandler
+ * The implementation delegates to {@link GetDataFrameOfElements} operation.
+ * Then the resulting {@link Dataset} of elements are split into two {@link Dataset}s
+ * based on the groups provided in the {@link uk.gov.gchq.gaffer.data.elementdefinition.view.View}.
+ * </p>
  */
 public class GetGraphFrameOfElementsHandler implements OutputOperationHandler<GetGraphFrameOfElements, GraphFrame> {
-
     @Override
     public GraphFrame doOperation(final GetGraphFrameOfElements operation, final Context context, final Store store) throws OperationException {
-        return doOperation(operation, context, (AccumuloStore) store);
-    }
-
-    public GraphFrame doOperation(final GetGraphFrameOfElements operation, final Context context,
-                                  final AccumuloStore store) throws OperationException {
-
         final GetDataFrameOfElements getDataFrame = new GetDataFrameOfElements.Builder()
                 .converters(operation.getConverters())
                 .view(operation.getView())
